Refractor end to end purchase test
This commit makes changes to the test case as suggested in code review.

The package name homepage was converted to lower case as per naming
conventions.

Page classes are now defined at the top of the test script rather than
in multiple sections repeatedly. This makes for better organization
and cleaner code.

WebDriverWait is utilized rather than time.sleep.

Changes as above will be applied to other test cases in this project
after approval.